requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
PyJWT
pydantic
passlib[bcrypt]
